Remove unneeded C-style casts
The `const void*` overload that was added now handles this.

+++ srcStatic/GameModules/IpDropDown.cpp
@@ -31,8 +31,8 @@ IPDropDown::IPDropDown()
 void IPDropDown::Load()
 {
 	// patch the call to EnableWindow so we can add strings.
-	Op2MemSetDword(populateComboBoxAddr, (int)&newEnableWindowAddr);
-	Op2MemSetDword(saveIpTextAddr, (int)&newInetAddr);
+	Op2MemSetDword(populateComboBoxAddr, &newEnableWindowAddr);
+	Op2MemSetDword(saveIpTextAddr, &newInetAddr);
 	Op2MemSet(nopDataAddr, 0x90, 14);
 }
